# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgroAI Smart Greenhouse API System",
    "description": "Backend API for AgroAI Smart Greenhouse API System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/greenhouses": {
      "get": {
        "tags": [
          "greenhouses"
        ],
        "summary": "List Greenhouses",
        "description": "List all greenhouses owned by the current user.",
        "operationId": "list_greenhouses_api_greenhouses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GreenhouseRead"
                  },
                  "type": "array",
                  "title": "Response List Greenhouses Api Greenhouses Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "greenhouses"
        ],
        "summary": "Create Greenhouse",
        "description": "Create a new greenhouse.",
        "operationId": "create_greenhouse_api_greenhouses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GreenhouseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenhouseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/greenhouses/{greenhouse_id}": {
      "get": {
        "tags": [
          "greenhouses"
        ],
        "summary": "Get Greenhouse",
        "operationId": "get_greenhouse_api_greenhouses__greenhouse_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenhouseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "greenhouses"
        ],
        "summary": "Edit Greenhouse",
        "operationId": "edit_greenhouse_api_greenhouses__greenhouse_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GreenhouseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenhouseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "greenhouses"
        ],
        "summary": "Delete Greenhouse",
        "operationId": "delete_greenhouse_api_greenhouses__greenhouse_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/greenhouses/{greenhouse_id}/telemetry": {
      "get": {
        "tags": [
          "greenhouses"
        ],
        "summary": "List Greenhouse Telemetry",
        "operationId": "list_greenhouse_telemetry_api_greenhouses__greenhouse_id__telemetry_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 720,
              "minimum": 1,
              "default": 24,
              "title": "Hours"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TelemetryRead"
                  },
                  "title": "Response List Greenhouse Telemetry Api Greenhouses  Greenhouse Id  Telemetry Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/greenhouses/{greenhouse_id}/commands/{command_id}": {
      "get": {
        "tags": [
          "greenhouses"
        ],
        "summary": "Get Device Command",
        "operationId": "get_device_command_api_greenhouses__greenhouse_id__commands__command_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Command Id"
            }
          },
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/greenhouses/{greenhouse_id}/ai/switch/{state}": {
      "post": {
        "tags": [
          "greenhouses"
        ],
        "summary": "Switch Mode Ai Control",
        "operationId": "switch_mode_ai_control_api_greenhouses__greenhouse_id__ai_switch__state__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "state",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "on",
                "off"
              ],
              "type": "string",
              "title": "State"
            }
          },
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "command_id": {
            "type": "string",
            "title": "Command Id"
          },
          "status": {
            "$ref": "#/components/schemas/CommandStatus"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "status"
        ],
        "title": "CommandResponse"
      },
      "CommandStatus": {
        "type": "string",
        "enum": [
          "pending",
          "published",
          "acknowledged",
          "failed",
          "timed_out"
        ],
        "title": "CommandStatus"
      },
...
      "GreenhouseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ai_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Mode"
          },
          "mqtt_topic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mqtt Topic Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GreenhouseCreate"
      },
      "GreenhouseRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ai_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Mode"
          },
          "mqtt_topic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mqtt Topic Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "stats": {
            "$ref": "#/components/schemas/GreenhouseStats"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "created_at",
          "stats"
        ],
        "title": "GreenhouseRead"
      },
...
      "GreenhouseUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "ai_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Mode"
          },
          "mqtt_topic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mqtt Topic Id"
          }
        },
        "type": "object",
        "title": "GreenhouseUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}